# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: RoseCloud IAM API
  version: 0.0.1
  description: Platform operator bootstrap/login, tenant invitations, tenant RBAC, and user session flows.
servers:
  - url: /api
paths:
...
  /sessions/login:
    post:
      operationId: userLogin
      summary: Login as an IAM user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserLoginRequest"
      responses:
        "200":
          description: User access token with refresh cookie
          headers:
            Set-Cookie:
              schema:
                type: string
              description: Refresh token cookie named rc_refresh
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/AccessTokenResponse"
                  - $ref: "#/components/schemas/FactorChallengeRequiredResponse"
        "401":
          description: Generic login rejection
        "429":
          description: Progressive cooldown after repeated failures for this email + client IP
          headers:
            Retry-After:
              schema:
                type: integer
              description: Seconds until another attempt may be accepted
  /sessions/refresh:
    post:
      operationId: userRefresh
      summary: Rotate the refresh cookie and mint a new user access token
      parameters:
        - in: cookie
          name: rc_refresh
          required: true
          schema:
            type: string
      responses:
        "200":
          description: New access token plus rotated refresh cookie
          headers:
            Set-Cookie:
              schema:
                type: string
              description: Rotated refresh token cookie named rc_refresh
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessTokenResponse"
        "401":
          description: Refresh token invalid, expired, or family revoked
        "409":
          description: Previous refresh token reused within grace window; client should retry with the winning cookie
  /sessions/logout:
    post:
      operationId: userLogout
      summary: Revoke the current login session family
      parameters:
        - in: cookie
          name: rc_refresh
          required: false
          schema:
            type: string
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Session family revoked and refresh cookie cleared
        "401":
          description: Missing refresh cookie and bearer token
...
  /sessions/factor-challenge:
    post:
      operationId: userFactorChallenge
      summary: Complete user FactorChallenge and issue session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FactorChallengeRequest"
      responses:
        "200":
          description: Access token with refresh cookie
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessTokenResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserLoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        totpCode:
          type: string
          pattern: '^\d{6}$'
      required: [email, password]
    AccessTokenResponse:
      type: object
      properties:
        accessToken:
          type: string
        tokenType:
          type: string
        expiresIn:
          type: integer
          format: int64
      required: [accessToken, tokenType, expiresIn]
...
    FactorChallengeRequest:
      type: object
      properties:
        challengeId:
          type: string
          format: uuid
        bindingId:
          type: string
        totpCode:
          type: string
        recoveryCode:
          type: string
      required: [challengeId]
    FactorChallengeRequiredResponse:
      type: object
      properties:
        status:
          type: string
          enum: [FACTOR_CHALLENGE_REQUIRED]
        challengeId:
          type: string
          format: uuid
        bindings:
          type: array
          items:
            $ref: "#/components/schemas/FactorBindingView"
      required: [status, challengeId, bindings]
    FactorBindingView:
      type: object
      properties:
        id:
          type: string
        kind:
          type: string
        softLabel:
          type: string
      required: [id, kind, softLabel]
...